web product analysis/requirement.txt
google-adk
python-dotenv
playwright